+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Next
 ----
 
  - add `object/matches`.
+ - add shorthand syntax to `array/*` methods.
 
 
 v0.3.0